refactor: refactor solve problem update tests
