+++ src/test/scala/satify/model/problems/ProblemWrongUseTest.scala
@@ -2,10 +2,6 @@ package satify.model.problems
 
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should.Matchers
-import satify.model.Result.*
-import satify.model.Solution
-import satify.update.solver.SolverType.DPLL
-import satify.update.solver.{Solver, SolverType}
 
 class ProblemWrongUseTest extends AnyFlatSpec with Matchers:
 
@@ -20,4 +16,3 @@ class ProblemWrongUseTest extends AnyFlatSpec with Matchers:
       NQueens(0)
     }
   }
-
