+++ src/test/scala/satify/update/SolveProblemUpdateTest.scala
@@ -2,41 +2,41 @@ package satify.update
 
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should.Matchers
-import satify.model.State
-import satify.model.errors.Error.InvalidInput
-import satify.model.problems.{NQueens, Problem}
+import satify.model.Result.{SAT, UNSAT}
+import satify.model.problems.{GraphColoring, NQueens, NurseScheduling}
+import satify.model.{Solution, State}
 import satify.update.Message.*
 import satify.update.Update.update
 
 class SolveProblemUpdateTest extends AnyFlatSpec with Matchers:
 
   val currentState: State = State()
-  val parameter: Int = 4
 
-  "When SolveProblem message containing a valid choice is sent to the update component it" should "return an updated State" in {
-    val problem: Problem = NQueens(4)
-    val newState = update(currentState, SolveProblem(problem))
-    import newState.*
-    input.isEmpty
-    && expression.isEmpty
-    && solution.isDefined
-    && cnf.isDefined
-    && error.isEmpty shouldBe true
+  "The solving of NQueens" should "return an updated State" in {
+    update(currentState, SolveProblem(NQueens(4))).solution should matchPattern { case Some(Solution(SAT, _, _)) =>
+    }
   }
 
-  "When SolveProblem message containing an invalid choice is sent to the update component it" should "return a defined Error" in {
-    val problem: Problem = NQueens(-2)
-    update(currentState, SolveProblem(problem)) shouldBe State(InvalidInput())
+  "Scheduling of 2 nurses, 2 days and 3 shifts" should "be UNSAT" in {
+    update(currentState, SolveProblem(NurseScheduling(2, 2, 3))).solution should matchPattern {
+      case Some(Solution(UNSAT, _, _)) =>
+    }
   }
 
-/*  "When Solve message containing a invalid expression is sent to the update component it" should "return a defined error" in {
-    val newState = update(currentState, Solve(invalidExpression))
-    import newState.*
-    input.isDefined
-      && expression.isEmpty
-      && solution.isEmpty
-      && cnf.isEmpty
-      && problem.isEmpty
-      && error.isDefined shouldBe true
+  "Scheduling of 3 nurses, 1 days and 3 shifts" should "be SAT" in {
+    update(currentState, SolveProblem(NurseScheduling(3, 1, 3))).solution should matchPattern {
+      case Some(Solution(SAT, _, _)) =>
+    }
   }
- */
+
+  "GraphColoring 3 linked nodes with 2 color and 1 auto-cycle" should "be UNSAT" in {
+    val problem = GraphColoring(List(("n1", "n2"), ("n2", "n3"), ("n2", "n2")), List("n1", "n2", "n3"), 2)
+    update(currentState, SolveProblem(problem)).solution should matchPattern { case Some(Solution(UNSAT, _, _)) =>
+    }
+  }
+
+  "GraphColoring 3 linked nodes with 2 colors" should "be SAT" in {
+    val problem = GraphColoring(List(("n1", "n2"), ("n2", "n3")), List("n1", "n2", "n3"), 2)
+    update(currentState, SolveProblem(problem)).solution should matchPattern { case Some(Solution(SAT, _, _)) =>
+    }
+  }
